work

+++ Test/Passes/InstructionSelection/RISCV64/load_invalid.mlir
@@ -3,8 +3,8 @@
 "builtin.module"() ({
     "func.func"()  <{function_type = (!llvm.ptr) -> ()}> ({
     ^bb0(%a: !llvm.ptr):
-        %val = "llvm.load"(%a) : (!llvm.ptr) -> i32
-        // CHECK: {{.*}} = "llvm.load"({{.*}}) <{"access_groups" = [], "alias_scopes" = [], "alignment" = 0 : i64, "noalias_scopes" = [], "tbaa" = []}> : (!llvm.ptr) -> i32
+        %val = "llvm.load"(%a) : (!llvm.ptr) -> i16
+        // CHECK: {{.*}} = "llvm.load"({{.*}}) <{"access_groups" = [], "alias_scopes" = [], "alignment" = 0 : i64, "noalias_scopes" = [], "tbaa" = []}> : (!llvm.ptr) -> i16
         "func.return"() : () -> ()
     }) : () -> ()
 }) : () -> ()
